Bug 707983: Fix buffer overflow in gsicc_set_device_profile_colorants
The function takes the string holding the plate names, and splits it into
individual names, storing them in a table.

But that table has a buid time maximum size, and the code wasn't checking that
the number of entries it pulled out of the string wasn't greater than that
limit.

This adds that check, as well as removing a pointless intermediate variable.