Bug 707967: Fix crash in gc caused by unitialised pointer
Changing the pdfwrite code to not use the "long" data type caused the struct
padding on 32 bit Windows to change for the pdf_pattern_t resource type.

The result was that the sizeof(pdf_pattern_t) was the same as the generic
sizeof(pdf_resource_t), and the code to intended to only initialise the
data not included in the pdf_resource_t no longer worked correctly.

The pdf_resource_t is relatively small, so just memset the entire object.