Bug 707937: Fix crash due to clist color handling.
The fix for bug 707830, involved us changing 'dev->num_std_colorants'
whenever we pushed/popped transparency groups.

Unfortunately, I'd missed that we also need to cope with this on
starting/stopping an smask. The starting case was already coped
with, but when we pop an smask we were leaving dev->num_std_colorants
set to the wrong value.

The fix is to restore the value here too.