Coverity ID #442392
The code here was actually correct, but it confuses Coverity because we
appear to be setting a strcture to NULL, but specifying a number of
bytes larger than said structure.

This is because the structure is embedded inside another structure, and
we are actually nulling many bytes of the enclosing structure, but
starting from the beginning of one of the enclosed structures.

This of course confuses Coverity (and, to be fair, me for a while)
but we must not NULL the entire large structure, because it is partially
filled in when we allocate it and we must not overwrite that.

However it looks like pdf_alloc_resource() has changed over the years
and it now actually NULLs the entire resource structure it has been
asked to allocate, so we no longer need to NULL it after allocation. So
remove the line that does that. It turns out there is one other place
where we do that (font_resource_alloc) so remove it from there too.