Bug 707937 continued: Fix SEGV with clist transparency handling.
In the previous fixes for this and bug 707830, I had modified
the pdf14_group_color_t to have a new field 'num_std_colorants'.
I was using that to store the PREVIOUS num_std_colorants state.

Turns out that this is different to the way all the other entries
in there are stored, where we store the CURRENT values.

When we pop a group, we look at the previous group_color field
rather than the one for the group we've just popped.

We fix the problems here by updating the usage to be in line with
the rest of the members, and ensuring that we set the value
properly in the base group_color entry.