pdfimg devices - fix math for calculating image scale
Bug #707944 "White border on page top when rendering a solid color PDF using the pdfimage24 device"

This was actually a customer report, and the customer cmae back to
support with additional information.

There are two problems in the original bug report, one is a bug, one
isn't.

When we render the original PDF to an image, 'clip' operations can mean
that, at some resolutions, there can be a 'gap' of up to 1 raster line
vertically, or 1 pixel horizontally which is left unmarked by the
clipped content when the user 'expected' it should butt up to some other
feature. The feature in this case is the edge of the media.

This isn't a bug, the problem can be reduced by rendering to a higher
resolution, then using DownScaleFactor to reduce the resolution back to
the target. For this file that means that the downscaler makes certain
that any such white pixel becomes half a pixel, and disappears.

The scond problem is in fact a bug. The scaling of the image used a
different calculation to the calculation of the media size, which
could result in the image not covering the entire media. In addition
the image scale factor was written (and calculated) as an integer, which
exacerbated the difference.

In this commit; use floats to write the scale factor, and use the same
calculation as the media size, so that the scale factor will always make
the image fit the media precisely.