pdfwrite device - Fix the OCR code
No bug for this, it was pointed out, in a private email, by a free user
while discussing fonts and PDF.

The new PDF interpreter completely breaks the OCR processing in
pdfwrite. This is because the code in gdevpdtt.c, ProcessTextForOCR()
assumed that the default text processing would return
TEXT_PROCESS_RENDER for every glyph that did not have an entry in the
glyph cache, in order to render the glyph.

The new code in pdfi doesn't need to do that, and so doesn't. This
means that we only stored the last character code and glyph, and not
until after the code in gdevpdfb.c, gdev_pdf_fill_mask() had used the
value, expecting it would be filled in, but these were always 0.

So here we take a pointer to the text enumerator and store it in the
device. Then instead of using the stored values in the device, in
gdev_pdf_fill_mask, we use the actual text enumerator and get the
values from there. This call is made within the default text processing
so the values are updated properly.

This seems to work correctly now for both PostScript and PDF input.

This still doesn't work properly though. The reason is that there was
a bug in gdevpdtw.c, pdf_simple_font_needs_ToUnicode() which was
fixed in commit 	8f62213019bc682eeb0ed9467d8841f3770cfda6

The fix means that Unicode code points that cannot be represented in
two bytes were previously passed through, but now they are not. The
return values from the OCR process are always returned as 4 bytes, so
we never use them after that commit.

So in OCRText(), if the Unicode code point is < 0xFFFF return it as a
2-byte code point, otherwise return it as a 4-byte code point. We still
can't use the 4 byte values, which means the font will not get a
ToUnicode CMap, but 2-byte cases will work.