pdfwrite devide - more work on 'clip' text rendering modes
Bug #708025 "incorrect clipping with text rendering mode 7 and additional clips in side gsaved states"

The problem is caused by commit f33ebdd329c424d5389d9e1aaaf84197d4e0b6fa
which was an enhancement (for the same customer) to deal with text
rendering modes involving clips as text, instead of writing out a long
path representing the shape of the text. Partly because when using
pdfwrite at low resolution (72 dpi) the resulting path is 'blocky' as it
is created at the resolution of the device, eg 72 dpi.

I hadn't appreciated that images, and possibly some other places, would
write out a clip, and if there was one active could end up doing a
grestore to get rid of it, before executing the operation. This leads to
the text clipping being elided in the output.

The fix is more complicated unfortunately. We capture the current
'bottom' of the saved gstate stack, then we update that bottom to point
at the state we have just created.

The 'put_clip' functions look at the difference between the current
depth of the saved gstate stack and the bottom before deciding whether
to emit a clip or a 'Q' to undo an existing clip. By altering the
'bottom' we can prevent them trying to grestore away the clip from the
text.

When we're done with the gstate holding the text clip we restore the
'bottom' value, and then restore back until we reach that point, which
removes the text clip.