pdfwrite - revert 	f33ebdd329c424d5389d9e1aaaf84197d4e0b6fa
The commit did as was intended, but there was a bug relating to clip
paths, text rendering mode 7, and gsave states. See bug 708025.

That bug was addressed with commit cc6d078f0901fcb2dd96047ed6333661c1816c33
but by then we had already started the release process fo 10.04.0. So
this commit reverts the original. Because it was an enhancement there
are no (visible) regressions and it is felt safer to do this than to
apply a fix which we have less confidence in and which has not been
significantly exercised.

The two commits remain on the master branch and will be included in the
next full release (currently expected to be 10.05.0).