PDF interpreter - don't gsave round clipping text modes
Bug #708013 "PDF to JPEG rendering issue box issue"

The basic problem is that the file is constructed in the usual insane
fashion. First it draw some text, then it draws an image which partially
obscures that text. Then it 'restores' the missing oprtion. It does that
by drawing the text again, except that it doesn't *draw* the text, it
uses rendering mode 7 to turn it into clip path. Ther problem occurs
because it sets the font and size inside a text block, *after* setting
rendering mode 7, and then exits the text block. The PDF file (rightly)
assumes that the font and size will persist. Because we gsave/grestore
round the text block, it doesn't.

It looks like I was too cautious when writing this bit of the
interpreter; there isn't any reason to preserve any extant path at the
time the clip starts because PDF doesn't let you do that (unlike, say
PostScript).

So simply remove the gsave/grestore pair around the operation.