PDF interpreter - check return values
Coverity IDs 443786, 443787

COmmit c8842ed68e80841a1e3ecd420b1df71dcbb54ee3 removed a couple of
gsvaes and a grestore that were'nt checking their return values. That
changed the overall balance of checked vs unchecked (in this file ?) and
caused Coverity to complain that these were not checked.

I'd have preferred the previous ones to be flagged as well but that
isn't hwo Coverity works.

Fix it by checking the return values for the gsaves and explicitly
casting the return value of a couple of grestores to void. Coverity
wasn't complaining about those but we may as well be clear.