PDF interpreter - Try images with invalid /Subtype as /Image subtype
Bug #708033 "Detect invalid image /Subtype"

As noted in the bug report, the PDF has an image with an /SMask to
provide an alpha channel. The SMask is an 8-bit gray image, it should
have a /subtype of /Image, in fact it is given as /Alpha, which is not
legal anywhere in PDF.

Acrobat, however opens the file, and we don't actually emit an error
message, we just silently ignore it.

This commit adds an error message when the /Subtype is unrecognised and
also checks PDFSTOPONERROR. If not set we try to treat the image or
form as an image. If the required entries are missing it'll fail anyway
(ie it's a Form or genuinely something we don't know how to handle).