PDF & PS interpreters - Fix /SeparationColorNames
Bug #707923 "GhostScript spot colors page managment"

The bug report claims the problem is with the PDF interpreter, but in
fact it is with the PostScript interpreter. The /SeparationColorNames
array is freed when the page is closed but because the dictionary
stored in the PostScript environment still contains the original
PostScript array, a call to setpagedevice does not resend the array.

Provided we don't send a setpagedevice, this doesn't seem to cause a
problem, though I think that's purely by coincidence; I think it should.

But the PDF interpreter calls setpagedevice on every page, so on the
second and subsequent pages, the SeparationColorNames array is lost. We
still produce the spot plates but the order is now different and the
halftone supplied in the bug report doesn't match the ink names with
the spot plate names, meaning the /Default gets used for all the spot
plates.

Note that exactly the same thing happens in PostScript, if the
setpagdevice operator is used; the fault is in the PostScript
setpagedevice craziness and the frankly bizarre way Ghostscript deals
with separations.

In any event, we can fix the problem by adding the /SeparationColorNames
key to the '.copiedkeys' array of keys which always get sent to the
device when setpagedevice is executed.

This did reveal an oddity in the PDF interpreter, caused by the error
handling work that was done recently, and I've chosen to add a fix for
that to this commit.