Utility scripts - fix ps2epsi
The ps2epsi scripts call ps2epsi.ps which uses selectdevice to change
between the eps2write and bit devices; but the script doesn't specify
these devices as permitted so they fail with an invalidaccess.

Specify in the script that the devices are permitted. Also update the
documentation a bit more to be clearer (with examples) about the
separator character for permitted devices.