pdfwrite - alter symbolic checking and Encoding emission if font not embedded
Bug #708047 "pdfwrite creates an invalid output file"

The problem occurs since we no longer embed fonts in the output file
which were not present in the input. We still check the font, in this
case the font we are substituting for the missing one, to see if we
need to write it as 'symbolic or not. The problem is that we check the
entire space of the font looking for glyphs which are not in Standard
Encoding, which for the majority of fonts will always find some
occurences.

That causes us to write the font as symbolic. If we're embedding the
font that isn't a problem, but if we aren't then Acrobat (and apparently
only Acrobat) gets upset and refuses to find a susbtitute.

So here, if we aren't embedding the font in the output, don't bother
checking the substituted font to see if we need to make it a symbolic
font.

Also, when deciding whether or not to embed an Encoding for TrueType fonts,
don't write one if we are not embedding the font, even if it is a
standard Encoding.

Cluster test shows diffs with some files, but no actual differences,
the change is just the flags and furniture in the file.