pdfwrite - use correct allocator when freeing patern states on cleanup
OSS-fuzz #368729562 (yes, new numbering scheme)

When closing the device we free any dangling pattern references (which
can be caused by broken patterns), but we were using the wrong
allocator, non-gc memory instead of regular memory. The pattern state
array is allocated in non-gc space, but the individual pattern states
held by it are allocated in gc-ed memory and are relocatable.