pdfwrite - address font substitution using TrueType Collections
bug #708048 " Wrong CIDFont subsetting"

The bug title isn't exactly illuminating. The problem here is that if
the input is a PDF file which does not have all the CIDFonts embedded,
and the user supplies a cidfmap which defines substitutions for the
non-embedded fonts as well as adding the font name to the AlwaysEmbed
array and, finally, if the defined substitute font is a TrueType
Collection, not simply a TrueType font, then the output PDF file has
multiple font subsets; one per glyph.

This is due to the fact that the new PDF interpreter presents the
substitute font in a different manner. The old interpreter, being
written in PostScript, presented the substitute type 2 CID font as
a TrueType font, the conversion to a font from a TTC being done
when the font is read. The new PDF interpreter presents the font
as a TrueType Collection.

There were a couple of places in the graphics library code relating
to copying fonts which only expected a simple TrueType font, and in
at least one place although there was code to deal with TrueType
Collections, it only ever copied subfontID 0.

We now carry round the subfontID as part of the gs_type42_data
structure, as well as the calculated offset to the actual tables
in the TTC that match the selected font. The code to emit the
font and check if two fonts are compatible can then use the offset
to directly access the tables of interest.