Bug 708009: Make pdfi_stream_to_buffer() more robust
Firstly, cope better with errors coming from streams due to invalid lengths
or broken filtered data.

Secondly, have pdfi_stream_to_buffer to use sgets rather than sfread, so even
when the reading of the data fails to complete, we can NULL the remainder
of the buffer.

Contributions from both myself and Ken Sharp in this commit.

Fixes several of the issues in the bug report.