graphics library - fix UMR
Bug #708009 cases 18 and 23

Commit e573968a29259c0cb72fc8a287ce744fe875be8a fixed a dangling
pointer by explicitly setting the 'b_tile' pointer to NULL, in the
'clues'.

But it only does it if the clue dev_color is a halftone. It seems that
image_init_clues only initialises all 256 clues if the image is an
8 bps image, for 1 bit none of the clues are initialised. This leads
to us testing uninitialised memory. This is actually entirely safe
because if the type is uninitialised we don't care about the rest of
the clue, it doesn't matter whether we NULL the b_tile or not.

But to stop valgrind having conniptions; don't bother testing the
type of the dev_color, just set all the b_tile pointers to NULL.