PDF interpreter - work around short colour palette string data
Bug #708055 "QR code not rendered to jpg from PDF"

The problem is the PDF file. It has a string data source for a
/Indexed CMYK colour space, the hival is given as 194, so the required
lookup table is 780 bytes (4 * 194). The supplied string is (after
unescaping) only 779 bytes long. Currently we throw an error on that
whcih trickles up and the interpreter continues, but doesn't draw the
image because the palette is broken.

This commit properly checks PDFSTOPONERROR and if that isn't set then
it sets the missing bytes of data to 0x00. Note we always use the longer
of the supplied string size, or the expected palette size, when
allocating the memory, so we won't overrun.