ps2write - allow embedding of base fonts
No bug report, noticed while working on other problems. There is
currently no way to embed the base 14 fonts in the output of ps2write,
because it falls into a clause intended for compatibility with old
versions of Acrobat Distiller, which never embedded the base 14.

This commit checks to see if we are producing PostScript output (so we
don't need to emulate an ancient Distiller) and if so then it checks the
/AlwaysEmbed list. If the font is present there then we embed it.