pdfwrite - fix Tr 7 text after image clip restoration
Bug #708081 "incorrect clipping with Tr mode 7 when a previous image involved a clip"

The problem was that we weren't checking to see if we had restored away
a clip, so that the clip in the output PDF graphcis state wasn't the
same as the clip in the input graphics state, before starting to write
out text in a rendering mode which involved clipping.

This led to the text clipi being combined with the previous clip, when
it should not have been.

Fix by simply checking to see if we need to restore back. The pdfwrite
function for checking this is misleadingly named :-(