pdfwrite - improve /AF handling in Catalog dictionary
Bug #708066 "Attachments missing in PDF but according to file size their data is embedded"

The attachments are not missing (as noted in the title), it is simply
that Acrobat can't find the embedded XML file. This is due to the
Catalog dictionary having two /AF entries, and a peculiar, though
reasonable requirement from Acrobat not to have an indirect object
be another indirect object.

Commit fc04328faeaa646359783eab980416020682ccaa updated the output of
PDF/A-3 files when writing EmbeddedFiles, to conform to the PDF/A-3
tech note 0010 regarding a /AF entry in the Catalog. This is now a
required entry (this is a clarification).

However, if a PostScript program also added a /AF key to the Catalog
dictionary then (because of the way COS object works) we would end up
writing a Catalog dictionary to the output file which contained *two*
AF keys. Not surprisingly this is confusing (not to mention illegal).

In addition, it seems that Acrobat doesn't like an indirect reference
which resolves to nothing more than another indirect reference.

This commit adds some simple detection for the existence of a /AF key
in the Catalog both when EmbeddedFiles are specified and when the
Catalog is emitted. If a PostScript program adds a /AF key to the
dictionary then the pdfwrite device no longer does so.

We also emit the indirect object diectly, rather than indirecting
through another indirect object, which satisfies Acrobat.

Finally I updated zugferd.ps not to emit a /AF itself, but let pdfwrite
do the work.