Fix compiler warnings
Robin pointed these out on Discord, in both cases the pointers just
need an appropriate cast.

We can't carry the OCR_enum as a pdf_text_enum_t pointer, because that
structure isn't defined at the point where it is included in the
pdfwrite device structure. Since it is derived from a gs_text_enum_t it
is safe to cast it to a gs_text_enum_t pointer, and back again.