pdfwrite - Fix /Line annotations with midifed CTM
Bug #708045 " Ghostscript PDFWRITE device flatten arrow annotation when Rotate page"

Practically the entire description in the bug thread is incorrect. The
annotation isn't 'flattened' (the usual use of this term would indicate
that the annotation was drawn as part of a content stream instead of
preserved as an annotation).

Annotations don't have 4 'edges' or 'corners'. They do have a Rect,
which is defined by 2 points given as x,y co-ordintates, but in this
case there is nothing wrong with the Rect.

The problem is, simply, that the /L key defines the start and end of
the Line annotation (again as 2 points) and we weren't updating the
values when the CTM was modified. This would affect any kind of
adjustment, including scaling, not just rotation.