Bug 708086: Fix problem with overprinting and transparency.
When taking account of overprinting while marking a rectangle
we first calculate the destination values without overprint.
We then only write the changed values back. To account for
changes in alpha, we need to scale the original values even
for supposedly unchanged values. Unfortunately, we were using
the new values rather than the old values to scale from!