Improve detection of X11 libraries
This hopefully fixes a long standing problem with not spotting that the core
X11 or X extensions libraries were missing.

Originally, missing X11 was not a fatal error (which was part of the problem).
Now, since the X11 devices are expected to be included, missing the X11 libs
will be a fatal error from configure, unless the user specifies
--without-x on the configure command line. The error messages explains that.