PDF interpreter - don't cache page dictionaries in the Pages tree
No bug report for this, file supplied by customer and may be
confidential.

The file has an 'unusual' layout. It has 9200+ pages, each page
dictionary has a Resources key whose value is an indrect reference. Each
page references the *same* Resources dictionary. So all the resources
required for all the pages are stored in a single dictionary. This
dictionary has ~20,000 elements.

This in itself isn't a problem. However, in order to retrieve the Page
dictionary for each page (which we need in order to run the page) we
walk the pages tree. Whenever we reach a leaf we read the relevant
Page dictionary. We then replace the indirect reference in the leaf
node with a pointer to the retrieved Page dictionary.

This is for performance reasons; should we need to rerun the page then
we don't need to read the dictionary again.

The problem with this is that the Page dictionary contains a copy of
the Resources dictionary, which is huge. If the Resources dictionary was
still in the cache then this wouldn't be a problem, as we'd just keep
a pointer to it. But the pages are complex enough that, by the time we
finish the page, the Resources dictionary has been flushed from the
cache. So when we read the next page dictionary we make a new copy of
the Resources dictionary for it.

Fairly quickly this exhausts memory.

Since the performance gain by keeping a copy of the page dictionary is
small, and rarely used (we'd need to run the same page multiple times)
we've decided to forgo this.

The commit here simply prevents the Page dictionary dereference code
from storing a copy of the Page dictionary in the Pages tree.

There is another small improvement I noticed in passing. The file has
Form XObjects which do not contain Resources dictionaries, but instead
inherit them from the Page dictionary. This is legal but the PDF
Reference notes it is not recommended. We were flagging this as an error
which it is not. Changed to a warning.