PDF interpreter - alter detection of PDF files
Bug #708115 " PDF files with comments before %PDF- generate unrecoverable error"

The problem is caused by the presence of what appear to be DSC comments
prior to the %PDF- identifier in the PDF file.

We specifically check for %! and %% before looking for %PDF- because of
AppleTalk PAP and the 'gv' application which apparently send such
sequences and expect an immediate response.

I don't think we need bother about AppleTalk any more, and testing with
the current version of gv (dated 2013) this commit doesn't seem to
cause any problems. Essentially we just allow the existing code which
looks for a %PDF- to fail, which then runs the input as PostScript
instead.

This may be a little slower, and its possible that my highly limited
testing is insufficient to uncover a real problem, so I've left the
code in place, but commented out, for now. If someone uncovers a real
problem then we'll have more information to help produce a better
solution.

Most of the change here is just white space.