pdfwrite - reverse the default for substitute font embedding
Bug #708112 "Letters from Danish alphabet are missing from the output"

Chris and I had discussed this previously. Commit
99b48f1992844e30ab4b05dbb2764ba99f7b66f4 altered the way that pdfwrite
deals with fonts which are substitutes for fonts which were used but
not included in the input file. We felt that the more 'correct'
behaviour was to not embed substitute fonts, because that produces an
output PDF file which is closer to the original PDF file.

After listening to feedback from users we've decided to reverse that
default. It is still possible to prevent substitute fonts being
embedded in the output in place of fonts missing from the input, but
the user must now set -dEmbedSubstituteFonts=false.

Nothing else changes, but this should be less surprising for users who
did not realise there had been a change. In particular, if
EmbedSubstituteFonts is false then any explicit substitutions will
need to be added to the AlwaysEmbed array, just as was the case in
the original 10.04.0 release.

Altered the documentation at the same time.