Implement special handling of CIDFallBack name in cidfmap
In the old PS implemented PDF interpreter, the name /CIDFallBack was handled
specially, allowing replacement of the default DroidSansFallback general
purpose CIDFont substitute (using the mapping AND not doing the Registry,
Ordering and Supplement validation we do on specific substitutions).

This adds the same special handling to the C implementation.