Fix bus error on ARM with tests_private/comparefiles/Bug690166.ps
The problem was a long string meant pdfwrite couldn't use its stack
based temporary buffer.

Allocating the temporary buffer as a string meant the buffer was byte
aligned, but the later code accessed the buffer as an array of gs_glyphs
which are an 8 byte type. Thus falling foul of the stricter than x86
alignment on the ARM architecture.