Bug 708125: Fix lack of halftoning with planar devices putimage.
Prior to the bugfix for 707767, a putimage to a planar device
would end up with the image data being sent as a normal image
with a simple colorspace. This would mean that the image data
was indeed halftoned, but that all the spot color information
would be lost, as it was transformed down to process colors.

The fix for 707767 meant that spot color stayed as spot color,
but the mechanism by which it was sent through
(fill_rectangle_hl_color) was incapable of halftoning, so the
halftoning was lost.

The fix implemented here is to amend the putimage fallback code
so that for devices that support devn, we'll use a DevN colorspace
instead of the simple colorspace. This means that data is properly
sent through and halftoned by the usual mechanisms.