PDF interpreter - increase uniqueness of objet definition names
Bug #708127 "pdfwrite corrupts hyperlink order when combining PDFs"

The problem is that in order to create a Link annotation in the
output file, from a Link annotation in the input file, which contains
a URL Action, we first need to create a URL action dictionary.

For historical reasons, we do this with a pdfmark. Because there is no
specific pdfmark for a URL Action, we create it as a general 'object'
and then refrence that object from the Link anntoation dictionary.

In order to reference an object, it must have a name. Obviously the
input file doesn't have a name (its PDF, not PostScript) so we have to
create one. The way this is done is to use the object number and
generation of the PDF Action dictionary. FOr a single input file this
is entirely reliable, because there can only be one object with a
given number and generation.

Unfortunately, as in this case, if there are multiple input files the
object number may not be unique. This leads to us creating a single
named object, which is referenced twice when we should have two named
objects.

This commit improves the uniqueness by doing a quick hash of the input
filename and appending that as another (32-bit) number. Files sent via
stdin are always copied to a temporary file, so there is *always* a
unique filename available, but we check for a NULL filename and just
use a fixed value, in case someone finds a way to break that.