pdfi: Fontmap/cidfmap - cope with cvn in mappings
In both Fontmap and cidfmap the key (font name) is expected to be a name object.

For some cases, it is easier to define the key as a string, and then convert it
to a name using the PostScript operator cvn. That's not an operator known to the
PDF interpreter and, as we leverage that for creating the font maps dicts in
pdfi, that will cause an error.

In fact, in pdfi, we automatically convert and string "key" to a name, so we can
simply replace the characters "cvn" with spaces in the buffer we're about to
interpret - this is viable because even a file with a large number of maps is
not that large.