PDF interpreter - don't corrupt error returns from page processing
No bug report, the file came from a customer.

At the end of processing a page, if we pushed a transparency
compositor (because the page uses transpanrecy) we need to pop it again.
However, if the pdfi_trans_end_group succeeded then the pop would
overwrite the return code from page processing, leading to us not
passing the error back up to the PostScript environment.

When PDFSTOPONERROR is true, we dhould reflect that error, not simply
swallow it.