Bug 708147: Implement default strip_tile_rect_devn.
In the example case, we're creating a pattern tile that is
too large to get into the tile cache.

In the non-hlcolor version, we fallback to default filling,
so the use of the tile is resolved pre-clist. We can't currently
do the same in the hlcolor version, as we don't have a default
strip_tile_rect_devn function.

So, implement one that works by calling fill_rectangle_hl_color.

The only problem here is that we don't have a graphics state
to pass in. This doesn't actually seem to be a problem in this
case though, as I can't see where it's ever used...

Maybe we'll need to update this in future, but it seems like a
reasonable first step.