pdfwrite - fix compression chooser
Bug report from a customer, but test file
tests/Ghent_V5.0/GWG161_Transp_Basic_BM_ICCBasedRGB_x4.pdf demonstrates
the issue as well.

Commit 3e54056d37c94deaea151e72451d8443495e60ad changed many uses of the
'long' data type because it is not the same size on all platforms.

Unfortunatley I mis-interpreted what the 'packed_data' member is used
for in the compression chooser, it's an ugly way of unpacking component
data from a buffer. Altering it to a 64-bit value changes the way it
behaves.

The unpacked data is scanned to try and detect the use of linework,
and if we think the image is linework then we don't use JPEG but instead
use Flate compression. Changing the data caused this to go awry and
select Flate intead of JPEG.

This doesn't affect many files at all (25 out of our 10000+ test suite)
and the visual effect of the error actually **improves** the quality
of the output (no JPEG artefacts), even though the output file is
slightly larger which is why I missed it. Even at that only 3 files
were more than 5% larger.

If we were not passing JPEG images through unchanged by default (in the
past they were decompressed and recompresed) this would have been much
more obvious.