pdfwrite - more fun with clip and text rendering mode 7
Bug #708151 "Problems with text rendering mode 7 and saved clip paths"

The pdfwrite device wasn't emitting the clip path before starting the
text for the clip rendering mode. This, obviously, made it impossible
to restore to that clip.

Fixing this is slightly awkward, because we need to mess about with the
stack of saved gstates, to ensure that we don't accidentally write out
the clip multiple times. In addition to emitting any pending clip before
we start the text, we also need to make sure we return the saved gstate
stack to the correct bottom and 'current' positions, which has meant
adding another variable to the device structure.

We also need access to the current graphcis state, in order to write the
current clip path, so the PDF interpreter needs to pass that in as a
parameter to the spec_op when preserving text rendering modes.
Previously this was just a flag, 0 or 1, now its either 0 (NULL) or a
pointer to the graphics state.