Bug 708152: Fix incomplete clist fallback.
As part of commit 597f97a5b7fe862377f7, the fix for bug 708147,
I tweaked the clist so that it would fallback to calling
the default implementation of strip_tile_rect_devn.

Unfortunately, I only did this for the first use in a tile, not
for all the vertical uses required. Fixed here.