PDF interpreter - consider /Rotate 180 with PDFFitPage and no rotation
Bug 708137 "Ghostscript 9.56.0 or above render incorrect result for a PDF page rotated 180 degree"

Realistically there is no 'incorrect' result. When you use PDFFitPage
the orientation of the output is whatever fits best, that's the way
it works.

However, when the content does not need to be rotated in order to fit
on the output we can consider a /Rotate of 180, and only that value
because other values would roate the content so that it no longer
fit on the output.

This commit simply does that.