PJL interpreter - alter the pdfmark processing
bug #708162 "Can't attach pdfmark via PJL"

Some time back I added code to allow PJL to process the PostScript
pdfmark and setdistillerparams operators. The reason for this is that
there was no way, in PCL, to set some options in the pdfwrite output.

As per this bug, the code bit-rotted. This was due to commit
ac6fa64af942ebc558cfd261ef39805db3f380e3

I'm not complettely certain what the purpose of this commit was, and
the commit logs don't shine much light on it either.

It looks to me like the intention is that 'resources', and I'm assuming
here we mean fonts, which are downloaded in the course of a PCL job
should be available to the next (PCL) job to come along.

Say for example I download MyLogoFont in the first job, then run another
job which uses the font but does not define it. Previously we'd have got
font substitution, with the relevant commit we would get the font
downloaded in the first job.

So far, so good. But....

The work is done by changing the type of reset performed in
pcl_impl_init_job from pcl_reset_initial to pcl_reset_printer. In order
to get the initial reset, a new reset is inserted into
pcl_impl_allocate_interp_instance.

So we used to get (from the bug report here) :

reset (pcl_reset_initial)
run a PJL job
reset (pcl_reset_initial)
run a PCL job

Now we get

reset (pcl_reset_initial)
reset(pcl_reset_printer)
run a PJL job
reset (pcl_reset_printer)
run a PCL job

The problem occurs because we only write the pdfmark when we get a
pcl_reset_initial. After we've run the PJL we never do a reset_initial
so we never emit the pdfmark. We can't emit it in the first reset,
because that happens before we've run the PJL.

I originally coded it this way because some pdfmarks can have incorrect
or surprising results if run multiple times, so I wanted to ensure they
would only be run once.

Sadly that's no longer possible, at least not without reverting the
commit which changed the reset type (and thereby reintroducing that bug)
or reworking the PCL interpreter to add another kind of reset, which is
far more work than is justified by this.

So I've chosen to emit the pdfmark on every printer reset. To be fair
that actually more closely follows the original code anyway. Given
that this is the first time in 3 years that anyone has raised this, it
can't be a commonly used feature, and this commit restores the actual
functionality, with a very small attendant risk that might cause some
unexpected behaviour with careless use of pdfmark.