pdfwrite - improve text clipping yet again
Working through the pdf-differences suite this cropped up because of
a potential change to 'd1' which I didn't actually commit. The same
problem could occur with a 'd0', however.

The problem is that we can get a type 3 font (for which clip modes do
not apply) which uses a type 1 font (for which they do). This led to
a mismatch with the graphics states.

This commit makes sure we don't start accumulating a clip when we have
already started accumulating a CharProc.

Also fix a typo in a comment from an earlier commit