PDF interpreter - alter default MissingWidth
This turned up in the Type 3 test in the 'pdf-differences' suite.

The type 3 font has no font descriptor, no 'space' glyph and no
/.notdef glpyh, but it uses the 'space' character code (32) and sets
Tw.

We had incorrectly set the MissingWidth of the font to 1000, which meant
that the space gained a width which it should not have had.

This also includes code to handle 'd1' type 3 CharProcs so that fill and
stroke operations use both the fill and stroke colours. We can't cache
the glyph bitmap in this case, so we handle it by treating like a 'd0'
case and rerendering each time it is used.

This caused problems, and in any case it seems from reading the spec
that the old behaviour (which matches Acrobat) might actully be correct
so it has been commented out. Left here for potential future use.