PCL interpreter - fix symbol set 19U table
Bug #708203 "Character U+00D0 LATIN CAPITAL LETTER ETH not printed using codeset Windows 1252 (19U)"

The code for symbol 0x00d0 appers to be incorrect. I've tracked it back
as far as commit 	79cec314a296c9a48049366596f83a0811428130 which claims
to be something to do with having all fonts accessed as Unicode instead
of 'MSL'.

MSL appears to mean 'Master Sybmol List'.

In any event, the symbol 0x00d0 was mapped to '0x110' at that time, and
does not appear to have been altered prior to that. So either it has
always been incorrect, or it was not updated with the commit above, and
should have been.

This commit simply changes the mapping to be the identity, which seems
to work.